+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/springdatajpa/FlashCardSDJpaService.java
@@ -7,7 +7,7 @@ import ru.dpopkov.knowthenics.repositories.FlashCardRepository;
 import ru.dpopkov.knowthenics.services.FlashCardService;
 
 @Service
-@Profile({"spring-data-jpa", "dev"})
+@Profile({"spring-data-jpa", "dev", "prod"})
 public class FlashCardSDJpaService extends AbstractSDJpaService<FlashCard> implements FlashCardService {
 
     public FlashCardSDJpaService(FlashCardRepository flashCardRepository) {