+++ knowthenics-data/src/main/java/ru/dpopkov/knowthenics/services/springdatajpa/TopicSDJpaService.java
@@ -7,7 +7,7 @@ import ru.dpopkov.knowthenics.repositories.TopicRepository;
 import ru.dpopkov.knowthenics.services.TopicService;
 
 @Service
-@Profile({"spring-data-jpa", "dev"})
+@Profile({"spring-data-jpa", "dev", "prod"})
 public class TopicSDJpaService extends AbstractSDJpaService<Topic> implements TopicService {
 
     public TopicSDJpaService(TopicRepository topicRepository) {